backend/requirements.txt
# Backend Requirements
# FastAPI Multi-Agent Policy Deliberation System

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WebSocket & Real-time
websockets==12.0
python-socketio==5.11.0
redis==5.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary==2.9.9  # Commented out - requires PostgreSQL installed. Use SQLite for dev.
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Existing Project Dependencies
# Using latest compatible versions
crewai>=0.203.1
langchain>=0.3.13
langchain-openai>=0.2.14
langchain-community>=0.3.13

# Explainability (SHAP/LIME)
shap==0.49.1
lime==0.2.0.1
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
scikit-learn==1.4.0

# Vector Store
pinecone==5.4.2

# Tools
wikipedia==1.4.0

# API Clients
google-search-results==2.4.2

# Utilities
tqdm==4.66.1
colorama==0.4.6
